+++ src/send.c
@@ -253,11 +253,33 @@ static int sendbufto_one_prepare_line(Client *to, char *msg)
 	len = strlen(p);
 	if (!len || (p[len - 1] != '\n'))
 	{
-		if (len > 510)
-			len = 510;
-		p[len++] = '\r';
-		p[len++] = '\n';
-		p[len] = '\0';
+		if (!IsServer(to) || !SupportBIGLINES(to->direction))
+		{
+			/* Normal case */
+			if (len > 510)
+				len = 510;
+			p[len++] = '\r';
+			p[len++] = '\n';
+			p[len] = '\0';
+		} else {
+			/* BIGLINES case:
+			 * - first 'if' is about the total line length,
+			 *   this is basically an optimized strlen(msg)
+			 * - the 'else' applies to non-mtags part,
+			 *   like in the 'Normal case' from above.
+			 */
+			if ((p - msg) + len > MAXLINELENGTH-3)
+			{
+				len = MAXLINELENGTH-3;
+				msg[len++] = '\r';
+				msg[len++] = '\n';
+				msg[len] = '\0';
+			} else {
+				p[len++] = '\r';
+				p[len++] = '\n';
+				p[len] = '\0';
+			}
+		}
 	}
 
 	/* Return length, that is: